Move cluster-defined enums/bitmaps into a separate header.
We want to put basic cluster-defined constants (enums/bitmaps) into a
header we can use from anywhere (just like data model concepts like
NodeId and whatnot).  But we don't want that for the C++ structs in
cluster-objects, which can have ABI issues.

Move the constants into a separate header, put that header into
src/lib/core in terms of include dependencies.
