+++ src/app/zap-templates/templates/app/cluster-enums.zapt
@@ -1,0 +1,46 @@
+{{> header}}
+
+#include <stdint.h>
+
+#pragma once
+
+namespace chip {
+namespace app {
+namespace Clusters {
+
+{{#zcl_clusters}}
+namespace {{asUpperCamelCase name}} {
+{{#zcl_enums}}
+
+{{#if (isWeaklyTypedEnum label)}}
+// Need to convert consumers to using the new enum classes, so we
+// don't just have casts all over.
+#ifdef CHIP_USE_ENUM_CLASS_FOR_IM_ENUM
+{{/if}}
+// Enum for {{label}}
+enum class {{asType label}} : {{asUnderlyingZclType type}} {
+{{#zcl_enum_items}}
+k{{asUpperCamelCase label}} = {{asHex value 2}},
+{{/zcl_enum_items}}
+};
+{{#if (isWeaklyTypedEnum label)}}
+#else // CHIP_USE_ENUM_CLASS_FOR_IM_ENUM
+using {{asType label}} = EmberAf{{asType label}};
+#endif
+{{/if}}
+{{/zcl_enums}}
+{{#zcl_bitmaps}}
+
+// Bitmap for {{label}}
+enum class {{asType label}} : {{asUnderlyingZclType type}} {
+{{#zcl_bitmap_items}}
+k{{asUpperCamelCase label}} = {{asHex mask}},
+{{/zcl_bitmap_items}}
+};
+{{/zcl_bitmaps}}
+} // namespace {{asUpperCamelCase name}}
+
+{{/zcl_clusters}}
+} // namespace Clusters
+} // namespace app
+} // namespace chip
