+++ src/lib/core/ClusterEnums.h
@@ -1,0 +1,28 @@
+/*
+ *
+ *    Copyright (c) 2022 Project CHIP Authors
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+#pragma once
+
+/**
+ * Enum and bitmap definitions from cluster specifications.
+ */
+
+// For now, include the generated enum file, since we are still working on
+// converting our enums to enum classes.  Once that is complete and things are
+// more stable, we can consider other options here.
+
+#include <app-common/zap-generated/cluster-enums.h>
